pyproject.toml
[project]
name = "move-me"
version = "0.1.0"
description = "Cross-platform screen locking tool with intelligent break management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psutil>=5.9.0",
    "plyer>=2.1",
    "pyyaml>=6.0",
    "typer>=0.9.0",
]

[project.scripts]
move-me = "move_me.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["move_me"]

[tool.uv.workspace]
members = [
    "temp",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]
